double -> BigDecimals, added CustomerAuthenticationManager, edited BankTerminal, Transaction

+++ app/src/core/TransactionOperations.java
@@ -1,6 +1,8 @@
 package AvaBank.core;
 
+import java.math.BigDecimal;
+
 public interface TransactionOperations {
-    public void depositAmount(double amount);
-    public void withdrawAmount(double amount);
+    public void depositAmount(BigDecimal amount);
+    public void withdrawAmount(BigDecimal amount);
 }
